Added support for items without description field

+++ pkg/data/data.go
@@ -48,7 +48,7 @@ func ToUnixTimestamp(s string) int64 {
 	return t.UnixMilli()
 }
 
-func ToInt(s string) int64 {
+func ToInt64(s string) int64 {
 	i, err := strconv.ParseInt(s, 10, 64)
 	if err != nil {
 		log.Panic(err)
